+++ tests/xml/test_port_interface.py
@@ -49,29 +49,31 @@ class TestInvalidationPolicy(unittest.TestCase):
 class TestSenderReceiverInterface(unittest.TestCase):
 
     def test_name_only(self):
-        element = ar_element.SenderReceiverInterface("ShortName")
+        element = ar_element.SenderReceiverInterface("InterfaceName")
         writer = autosar.xml.Writer()
         xml = '''<SENDER-RECEIVER-INTERFACE>
-  <SHORT-NAME>ShortName</SHORT-NAME>
+  <SHORT-NAME>InterfaceName</SHORT-NAME>
 </SENDER-RECEIVER-INTERFACE>'''
         self.assertEqual(writer.write_str_elem(element), xml)
         reader = autosar.xml.Reader()
         elem: ar_element.SenderReceiverInterface = reader.read_str_elem(xml)
         self.assertIsInstance(elem, ar_element.SenderReceiverInterface)
-        self.assertEqual(elem.name, "ShortName")
+        self.assertEqual(elem.name, "InterfaceName")
 
-    def test_make_element_with_ref_to_impl_type(self):
-        ref_to_impl_type = ar_element.AutosarDataTypeRef("DataTypes/OnOff_T",
-                                                         ar_enum.IdentifiableSubTypes.IMPLEMENTATION_DATA_TYPE)
-        element = ar_element.SenderReceiverInterface("InterfaceName")
-        element.make_data_element("Element1", type_ref=ref_to_impl_type)
+    def test_create_interface_with_one_element_using_constructor(self):
+        ref_to_uint8_type = ar_element.AutosarDataTypeRef("AUTOSAR_Platform/ImplementationDataTypes/uint8",
+                                                          ar_enum.IdentifiableSubTypes.IMPLEMENTATION_DATA_TYPE)
+        element = ar_element.SenderReceiverInterface("InterfaceName",
+                                                     ar_element.VariableDataPrototype("Element1",
+                                                                                      type_ref=ref_to_uint8_type))
+
         writer = autosar.xml.Writer()
         xml = '''<SENDER-RECEIVER-INTERFACE>
   <SHORT-NAME>InterfaceName</SHORT-NAME>
   <DATA-ELEMENTS>
     <VARIABLE-DATA-PROTOTYPE>
       <SHORT-NAME>Element1</SHORT-NAME>
-      <TYPE-TREF DEST="IMPLEMENTATION-DATA-TYPE">DataTypes/OnOff_T</TYPE-TREF>
+      <TYPE-TREF DEST="IMPLEMENTATION-DATA-TYPE">AUTOSAR_Platform/ImplementationDataTypes/uint8</TYPE-TREF>
     </VARIABLE-DATA-PROTOTYPE>
   </DATA-ELEMENTS>
 </SENDER-RECEIVER-INTERFACE>'''
@@ -83,27 +85,28 @@ class TestSenderReceiverInterface(unittest.TestCase):
         self.assertEqual(len(elem.data_elements), 1)
         data_element = elem.data_elements[0]
         self.assertEqual(data_element.name, "Element1")
-        self.assertEqual(str(data_element.type_ref), "DataTypes/OnOff_T")
+        self.assertEqual(str(data_element.type_ref), "AUTOSAR_Platform/ImplementationDataTypes/uint8")
 
-    def test_make_two_elements_with_ref_to_impl_type(self):
-        ref_to_impl_type1 = ar_element.AutosarDataTypeRef("DataTypes/OnOff_T",
+    def test_create_interface_with_two_elements_using_constructor(self):
+        ref_to_uint8_type = ar_element.AutosarDataTypeRef("AUTOSAR_Platform/ImplementationDataTypes/uint8",
                                                           ar_enum.IdentifiableSubTypes.IMPLEMENTATION_DATA_TYPE)
-        ref_to_impl_type2 = ar_element.AutosarDataTypeRef("DataTypes/uint8",
-                                                          ar_enum.IdentifiableSubTypes.IMPLEMENTATION_DATA_TYPE)
-        element = ar_element.SenderReceiverInterface("InterfaceName")
-        element.make_data_element("Element1", type_ref=ref_to_impl_type1)
-        element.make_data_element("Element2", type_ref=ref_to_impl_type2)
+        ref_to_uint16_type = ar_element.AutosarDataTypeRef("AUTOSAR_Platform/ImplementationDataTypes/uint16",
+                                                           ar_enum.IdentifiableSubTypes.IMPLEMENTATION_DATA_TYPE)
+        element = ar_element.SenderReceiverInterface("InterfaceName", [
+            ar_element.VariableDataPrototype("Element1", type_ref=ref_to_uint8_type),
+            ar_element.VariableDataPrototype("Element2", type_ref=ref_to_uint16_type)
+        ])
         writer = autosar.xml.Writer()
         xml = '''<SENDER-RECEIVER-INTERFACE>
   <SHORT-NAME>InterfaceName</SHORT-NAME>
   <DATA-ELEMENTS>
     <VARIABLE-DATA-PROTOTYPE>
       <SHORT-NAME>Element1</SHORT-NAME>
-      <TYPE-TREF DEST="IMPLEMENTATION-DATA-TYPE">DataTypes/OnOff_T</TYPE-TREF>
+      <TYPE-TREF DEST="IMPLEMENTATION-DATA-TYPE">AUTOSAR_Platform/ImplementationDataTypes/uint8</TYPE-TREF>
     </VARIABLE-DATA-PROTOTYPE>
     <VARIABLE-DATA-PROTOTYPE>
       <SHORT-NAME>Element2</SHORT-NAME>
-      <TYPE-TREF DEST="IMPLEMENTATION-DATA-TYPE">DataTypes/uint8</TYPE-TREF>
+      <TYPE-TREF DEST="IMPLEMENTATION-DATA-TYPE">AUTOSAR_Platform/ImplementationDataTypes/uint16</TYPE-TREF>
     </VARIABLE-DATA-PROTOTYPE>
   </DATA-ELEMENTS>
 </SENDER-RECEIVER-INTERFACE>'''
@@ -115,17 +118,76 @@ class TestSenderReceiverInterface(unittest.TestCase):
         self.assertEqual(len(elem.data_elements), 2)
         data_element = elem.data_elements[0]
         self.assertEqual(data_element.name, "Element1")
-        self.assertEqual(str(data_element.type_ref), "DataTypes/OnOff_T")
+        self.assertEqual(str(data_element.type_ref), "AUTOSAR_Platform/ImplementationDataTypes/uint8")
         data_element = elem.data_elements[1]
         self.assertEqual(data_element.name, "Element2")
-        self.assertEqual(str(data_element.type_ref), "DataTypes/uint8")
+        self.assertEqual(str(data_element.type_ref), "AUTOSAR_Platform/ImplementationDataTypes/uint16")
+
+    def test_create_interface_with_one_element_using_make_method(self):
+        ref_to_uint8_type = ar_element.AutosarDataTypeRef("AUTOSAR_Platform/ImplementationDataTypes/uint8",
+                                                          ar_enum.IdentifiableSubTypes.IMPLEMENTATION_DATA_TYPE)
+        element = ar_element.SenderReceiverInterface("InterfaceName")
+        element.make_data_element("Element1", type_ref=ref_to_uint8_type)
+        writer = autosar.xml.Writer()
+        xml = '''<SENDER-RECEIVER-INTERFACE>
+  <SHORT-NAME>InterfaceName</SHORT-NAME>
+  <DATA-ELEMENTS>
+    <VARIABLE-DATA-PROTOTYPE>
+      <SHORT-NAME>Element1</SHORT-NAME>
+      <TYPE-TREF DEST="IMPLEMENTATION-DATA-TYPE">AUTOSAR_Platform/ImplementationDataTypes/uint8</TYPE-TREF>
+    </VARIABLE-DATA-PROTOTYPE>
+  </DATA-ELEMENTS>
+</SENDER-RECEIVER-INTERFACE>'''
+        self.assertEqual(writer.write_str_elem(element), xml)
+        reader = autosar.xml.Reader()
+        elem: ar_element.SenderReceiverInterface = reader.read_str_elem(xml)
+        self.assertIsInstance(elem, ar_element.SenderReceiverInterface)
+        self.assertEqual(elem.name, "InterfaceName")
+        self.assertEqual(len(elem.data_elements), 1)
+        data_element = elem.data_elements[0]
+        self.assertEqual(data_element.name, "Element1")
+        self.assertEqual(str(data_element.type_ref), "AUTOSAR_Platform/ImplementationDataTypes/uint8")
+
+    def test_create_interface_with_two_elements_using_make_method(self):
+        ref_to_uint8_type = ar_element.AutosarDataTypeRef("AUTOSAR_Platform/ImplementationDataTypes/uint8",
+                                                          ar_enum.IdentifiableSubTypes.IMPLEMENTATION_DATA_TYPE)
+        ref_to_uint16_type = ar_element.AutosarDataTypeRef("AUTOSAR_Platform/ImplementationDataTypes/uint16",
+                                                           ar_enum.IdentifiableSubTypes.IMPLEMENTATION_DATA_TYPE)
+        element = ar_element.SenderReceiverInterface("InterfaceName")
+        element.make_data_element("Element1", type_ref=ref_to_uint8_type)
+        element.make_data_element("Element2", type_ref=ref_to_uint16_type)
+        writer = autosar.xml.Writer()
+        xml = '''<SENDER-RECEIVER-INTERFACE>
+  <SHORT-NAME>InterfaceName</SHORT-NAME>
+  <DATA-ELEMENTS>
+    <VARIABLE-DATA-PROTOTYPE>
+      <SHORT-NAME>Element1</SHORT-NAME>
+      <TYPE-TREF DEST="IMPLEMENTATION-DATA-TYPE">AUTOSAR_Platform/ImplementationDataTypes/uint8</TYPE-TREF>
+    </VARIABLE-DATA-PROTOTYPE>
+    <VARIABLE-DATA-PROTOTYPE>
+      <SHORT-NAME>Element2</SHORT-NAME>
+      <TYPE-TREF DEST="IMPLEMENTATION-DATA-TYPE">AUTOSAR_Platform/ImplementationDataTypes/uint16</TYPE-TREF>
+    </VARIABLE-DATA-PROTOTYPE>
+  </DATA-ELEMENTS>
+</SENDER-RECEIVER-INTERFACE>'''
+        self.assertEqual(writer.write_str_elem(element), xml)
+        reader = autosar.xml.Reader()
+        elem: ar_element.SenderReceiverInterface = reader.read_str_elem(xml)
+        self.assertIsInstance(elem, ar_element.SenderReceiverInterface)
+        self.assertEqual(elem.name, "InterfaceName")
+        self.assertEqual(len(elem.data_elements), 2)
+        data_element = elem.data_elements[0]
+        self.assertEqual(data_element.name, "Element1")
+        self.assertEqual(str(data_element.type_ref), "AUTOSAR_Platform/ImplementationDataTypes/uint8")
+        data_element = elem.data_elements[1]
+        self.assertEqual(data_element.name, "Element2")
+        self.assertEqual(str(data_element.type_ref), "AUTOSAR_Platform/ImplementationDataTypes/uint16")
 
     def test_make_element_with_invalidation_policy(self):
-        ref_to_impl_type = ar_element.AutosarDataTypeRef("DataTypes/OnOff_T",
-                                                         ar_enum.IdentifiableSubTypes.IMPLEMENTATION_DATA_TYPE)
+        ref_to_uint8_type = ar_element.AutosarDataTypeRef("AUTOSAR_Platform/ImplementationDataTypes/uint8",
+                                                          ar_enum.IdentifiableSubTypes.IMPLEMENTATION_DATA_TYPE)
         element = ar_element.SenderReceiverInterface("InterfaceName")
-        element.make_data_element("Element1",
-                                  type_ref=ref_to_impl_type)
+        element.make_data_element("Element1", type_ref=ref_to_uint8_type)
         element.make_invalidation_policy("/PortInterfaces/InterfaceName/Element1",
                                          handle_invalid=ar_enum.HandleInvalid.REPLACE)
         writer = autosar.xml.Writer()
@@ -134,7 +196,7 @@ class TestSenderReceiverInterface(unittest.TestCase):
   <DATA-ELEMENTS>
     <VARIABLE-DATA-PROTOTYPE>
       <SHORT-NAME>Element1</SHORT-NAME>
-      <TYPE-TREF DEST="IMPLEMENTATION-DATA-TYPE">DataTypes/OnOff_T</TYPE-TREF>
+      <TYPE-TREF DEST="IMPLEMENTATION-DATA-TYPE">AUTOSAR_Platform/ImplementationDataTypes/uint8</TYPE-TREF>
     </VARIABLE-DATA-PROTOTYPE>
   </DATA-ELEMENTS>
   <INVALIDATION-POLICYS>
@@ -155,6 +217,7 @@ class TestSenderReceiverInterface(unittest.TestCase):
         self.assertEqual(len(elem.invalidation_policies), 1)
         invalidataion_policy = elem.invalidation_policies[0]
         self.assertEqual(str(invalidataion_policy.data_element_ref), "/PortInterfaces/InterfaceName/Element1")
+        self.assertEqual(invalidataion_policy.handle_invalid, ar_enum.HandleInvalid.REPLACE)
 
     def test_make_element_with_numerical_init_value(self):
         ref_to_impl_type = ar_element.AutosarDataTypeRef("DataTypes/OnOff_T",
@@ -191,7 +254,32 @@ class TestSenderReceiverInterface(unittest.TestCase):
         self.assertIsInstance(init_value, ar_element.NumericalValueSpecification)
         self.assertEqual(init_value.value, 3)
 
-    def test_make_service_interface(self):
+    def test_create_service_interface_with_one_element_using_constructor(self):
+        ref_to_impl_type = ar_element.AutosarDataTypeRef("DataTypes/OnOff_T",
+                                                         ar_enum.IdentifiableSubTypes.IMPLEMENTATION_DATA_TYPE)
+        element = ar_element.SenderReceiverInterface("ServiceInterface",
+                                                     ar_element.VariableDataPrototype("Element1",
+                                                                                      type_ref=ref_to_impl_type),
+                                                     is_service=True)
+        writer = autosar.xml.Writer()
+        xml = '''<SENDER-RECEIVER-INTERFACE>
+  <SHORT-NAME>ServiceInterface</SHORT-NAME>
+  <IS-SERVICE>true</IS-SERVICE>
+  <DATA-ELEMENTS>
+    <VARIABLE-DATA-PROTOTYPE>
+      <SHORT-NAME>Element1</SHORT-NAME>
+      <TYPE-TREF DEST="IMPLEMENTATION-DATA-TYPE">DataTypes/OnOff_T</TYPE-TREF>
+    </VARIABLE-DATA-PROTOTYPE>
+  </DATA-ELEMENTS>
+</SENDER-RECEIVER-INTERFACE>'''
+        self.assertEqual(writer.write_str_elem(element), xml)
+        reader = autosar.xml.Reader()
+        elem: ar_element.SenderReceiverInterface = reader.read_str_elem(xml)
+        self.assertIsInstance(elem, ar_element.SenderReceiverInterface)
+        self.assertEqual(elem.name, "ServiceInterface")
+        self.assertTrue(elem.is_service)
+
+    def test_create_service_interface_with_one_element_using_make_method(self):
         ref_to_impl_type = ar_element.AutosarDataTypeRef("DataTypes/OnOff_T",
                                                          ar_enum.IdentifiableSubTypes.IMPLEMENTATION_DATA_TYPE)
         element = ar_element.SenderReceiverInterface("ServiceInterface", is_service=True)
@@ -213,6 +301,280 @@ class TestSenderReceiverInterface(unittest.TestCase):
         self.assertIsInstance(elem, ar_element.SenderReceiverInterface)
         self.assertEqual(elem.name, "ServiceInterface")
         self.assertTrue(elem.is_service)
+
+
+class TestNvDataInterface(unittest.TestCase):
+
+    def test_name_only(self):
+        element = ar_element.NvDataInterface("InterfaceName")
+        writer = autosar.xml.Writer()
+        xml = '''<NV-DATA-INTERFACE>
+  <SHORT-NAME>InterfaceName</SHORT-NAME>
+</NV-DATA-INTERFACE>'''
+        self.assertEqual(writer.write_str_elem(element), xml)
+        reader = autosar.xml.Reader()
+        elem: ar_element.NvDataInterface = reader.read_str_elem(xml)
+        self.assertIsInstance(elem, ar_element.NvDataInterface)
+        self.assertEqual(elem.name, "InterfaceName")
+
+    def test_create_interface_with_one_element_using_constructor(self):
+        ref_to_uint8_type = ar_element.AutosarDataTypeRef("AUTOSAR_Platform/ImplementationDataTypes/uint8",
+                                                          ar_enum.IdentifiableSubTypes.IMPLEMENTATION_DATA_TYPE)
+        element = ar_element.NvDataInterface("InterfaceName",
+                                             ar_element.VariableDataPrototype("Data1", type_ref=ref_to_uint8_type))
+        writer = autosar.xml.Writer()
+        xml = '''<NV-DATA-INTERFACE>
+  <SHORT-NAME>InterfaceName</SHORT-NAME>
+  <NV-DATAS>
+    <VARIABLE-DATA-PROTOTYPE>
+      <SHORT-NAME>Data1</SHORT-NAME>
+      <TYPE-TREF DEST="IMPLEMENTATION-DATA-TYPE">AUTOSAR_Platform/ImplementationDataTypes/uint8</TYPE-TREF>
+    </VARIABLE-DATA-PROTOTYPE>
+  </NV-DATAS>
+</NV-DATA-INTERFACE>'''
+        self.assertEqual(writer.write_str_elem(element), xml)
+        reader = autosar.xml.Reader()
+        elem: ar_element.NvDataInterface = reader.read_str_elem(xml)
+        self.assertIsInstance(elem, ar_element.NvDataInterface)
+        self.assertEqual(elem.name, "InterfaceName")
+        self.assertEqual(len(elem.data_elements), 1)
+        data_element = elem.data_elements[0]
+        self.assertEqual(data_element.name, "Data1")
+        self.assertEqual(str(data_element.type_ref), "AUTOSAR_Platform/ImplementationDataTypes/uint8")
+
+    def test_create_interface_with_two_elements_using_constructor(self):
+        ref_to_uint8_type = ar_element.AutosarDataTypeRef("AUTOSAR_Platform/ImplementationDataTypes/uint8",
+                                                          ar_enum.IdentifiableSubTypes.IMPLEMENTATION_DATA_TYPE)
+        ref_to_uint16_type = ar_element.AutosarDataTypeRef("AUTOSAR_Platform/ImplementationDataTypes/uint16",
+                                                           ar_enum.IdentifiableSubTypes.IMPLEMENTATION_DATA_TYPE)
+        element = ar_element.NvDataInterface("InterfaceName",
+                                             [ar_element.VariableDataPrototype("Data1", type_ref=ref_to_uint8_type),
+                                              ar_element.VariableDataPrototype("Data2", type_ref=ref_to_uint16_type)
+                                              ])
+        writer = autosar.xml.Writer()
+        xml = '''<NV-DATA-INTERFACE>
+  <SHORT-NAME>InterfaceName</SHORT-NAME>
+  <NV-DATAS>
+    <VARIABLE-DATA-PROTOTYPE>
+      <SHORT-NAME>Data1</SHORT-NAME>
+      <TYPE-TREF DEST="IMPLEMENTATION-DATA-TYPE">AUTOSAR_Platform/ImplementationDataTypes/uint8</TYPE-TREF>
+    </VARIABLE-DATA-PROTOTYPE>
+    <VARIABLE-DATA-PROTOTYPE>
+      <SHORT-NAME>Data2</SHORT-NAME>
+      <TYPE-TREF DEST="IMPLEMENTATION-DATA-TYPE">AUTOSAR_Platform/ImplementationDataTypes/uint16</TYPE-TREF>
+    </VARIABLE-DATA-PROTOTYPE>
+  </NV-DATAS>
+</NV-DATA-INTERFACE>'''
+        self.assertEqual(writer.write_str_elem(element), xml)
+        reader = autosar.xml.Reader()
+        elem: ar_element.NvDataInterface = reader.read_str_elem(xml)
+        self.assertIsInstance(elem, ar_element.NvDataInterface)
+        self.assertEqual(elem.name, "InterfaceName")
+        self.assertEqual(len(elem.data_elements), 2)
+        data_element = elem.data_elements[0]
+        self.assertEqual(data_element.name, "Data1")
+        self.assertEqual(str(data_element.type_ref), "AUTOSAR_Platform/ImplementationDataTypes/uint8")
+        data_element = elem.data_elements[1]
+        self.assertEqual(data_element.name, "Data2")
+        self.assertEqual(str(data_element.type_ref), "AUTOSAR_Platform/ImplementationDataTypes/uint16")
+
+    def test_create_interface_with_one_element_using_make_method(self):
+        ref_to_uint8_type = ar_element.AutosarDataTypeRef("AUTOSAR_Platform/ImplementationDataTypes/uint8",
+                                                          ar_enum.IdentifiableSubTypes.IMPLEMENTATION_DATA_TYPE)
+        element = ar_element.NvDataInterface("InterfaceName")
+        element.make_data_element("Data1", type_ref=ref_to_uint8_type)
+        writer = autosar.xml.Writer()
+        xml = '''<NV-DATA-INTERFACE>
+  <SHORT-NAME>InterfaceName</SHORT-NAME>
+  <NV-DATAS>
+    <VARIABLE-DATA-PROTOTYPE>
+      <SHORT-NAME>Data1</SHORT-NAME>
+      <TYPE-TREF DEST="IMPLEMENTATION-DATA-TYPE">AUTOSAR_Platform/ImplementationDataTypes/uint8</TYPE-TREF>
+    </VARIABLE-DATA-PROTOTYPE>
+  </NV-DATAS>
+</NV-DATA-INTERFACE>'''
+        self.assertEqual(writer.write_str_elem(element), xml)
+        reader = autosar.xml.Reader()
+        elem: ar_element.NvDataInterface = reader.read_str_elem(xml)
+        self.assertIsInstance(elem, ar_element.NvDataInterface)
+        self.assertEqual(elem.name, "InterfaceName")
+        self.assertEqual(len(elem.data_elements), 1)
+        data_element = elem.data_elements[0]
+        self.assertEqual(data_element.name, "Data1")
+        self.assertEqual(str(data_element.type_ref), "AUTOSAR_Platform/ImplementationDataTypes/uint8")
+
+    def test_create_interface_with_two_elements_using_make_method(self):
+        ref_to_uint8_type = ar_element.AutosarDataTypeRef("AUTOSAR_Platform/ImplementationDataTypes/uint8",
+                                                          ar_enum.IdentifiableSubTypes.IMPLEMENTATION_DATA_TYPE)
+        ref_to_uint16_type = ar_element.AutosarDataTypeRef("AUTOSAR_Platform/ImplementationDataTypes/uint16",
+                                                           ar_enum.IdentifiableSubTypes.IMPLEMENTATION_DATA_TYPE)
+        element = ar_element.NvDataInterface("InterfaceName")
+        element.make_data_element("Data1", type_ref=ref_to_uint8_type)
+        element.make_data_element("Data2", type_ref=ref_to_uint16_type)
+
+        writer = autosar.xml.Writer()
+        xml = '''<NV-DATA-INTERFACE>
+  <SHORT-NAME>InterfaceName</SHORT-NAME>
+  <NV-DATAS>
+    <VARIABLE-DATA-PROTOTYPE>
+      <SHORT-NAME>Data1</SHORT-NAME>
+      <TYPE-TREF DEST="IMPLEMENTATION-DATA-TYPE">AUTOSAR_Platform/ImplementationDataTypes/uint8</TYPE-TREF>
+    </VARIABLE-DATA-PROTOTYPE>
+    <VARIABLE-DATA-PROTOTYPE>
+      <SHORT-NAME>Data2</SHORT-NAME>
+      <TYPE-TREF DEST="IMPLEMENTATION-DATA-TYPE">AUTOSAR_Platform/ImplementationDataTypes/uint16</TYPE-TREF>
+    </VARIABLE-DATA-PROTOTYPE>
+  </NV-DATAS>
+</NV-DATA-INTERFACE>'''
+        self.assertEqual(writer.write_str_elem(element), xml)
+        reader = autosar.xml.Reader()
+        elem: ar_element.NvDataInterface = reader.read_str_elem(xml)
+        self.assertIsInstance(elem, ar_element.NvDataInterface)
+        self.assertEqual(elem.name, "InterfaceName")
+        self.assertEqual(len(elem.data_elements), 2)
+        data_element = elem.data_elements[0]
+        self.assertEqual(data_element.name, "Data1")
+        self.assertEqual(str(data_element.type_ref), "AUTOSAR_Platform/ImplementationDataTypes/uint8")
+        data_element = elem.data_elements[1]
+        self.assertEqual(data_element.name, "Data2")
+        self.assertEqual(str(data_element.type_ref), "AUTOSAR_Platform/ImplementationDataTypes/uint16")
+
+
+class TestParameterInterface(unittest.TestCase):
+
+    def test_name_only(self):
+        element = ar_element.ParameterInterface("InterfaceName")
+        writer = autosar.xml.Writer()
+        xml = '''<PARAMETER-INTERFACE>
+  <SHORT-NAME>InterfaceName</SHORT-NAME>
+</PARAMETER-INTERFACE>'''
+        self.assertEqual(writer.write_str_elem(element), xml)
+        reader = autosar.xml.Reader()
+        elem: ar_element.ParameterInterface = reader.read_str_elem(xml)
+        self.assertIsInstance(elem, ar_element.ParameterInterface)
+        self.assertEqual(elem.name, "InterfaceName")
+
+    def test_create_interface_with_one_parameter_using_constructor(self):
+        ref_to_uint8_type = ar_element.AutosarDataTypeRef("AUTOSAR_Platform/ImplementationDataTypes/uint8",
+                                                          ar_enum.IdentifiableSubTypes.IMPLEMENTATION_DATA_TYPE)
+        element = ar_element.ParameterInterface("InterfaceName",
+                                                ar_element.ParameterDataPrototype("Param1", type_ref=ref_to_uint8_type))
+        writer = autosar.xml.Writer()
+        xml = '''<PARAMETER-INTERFACE>
+  <SHORT-NAME>InterfaceName</SHORT-NAME>
+  <PARAMETERS>
+    <PARAMETER-DATA-PROTOTYPE>
+      <SHORT-NAME>Param1</SHORT-NAME>
+      <TYPE-TREF DEST="IMPLEMENTATION-DATA-TYPE">AUTOSAR_Platform/ImplementationDataTypes/uint8</TYPE-TREF>
+    </PARAMETER-DATA-PROTOTYPE>
+  </PARAMETERS>
+</PARAMETER-INTERFACE>'''
+        self.assertEqual(writer.write_str_elem(element), xml)
+        reader = autosar.xml.Reader()
+        elem: ar_element.SenderReceiverInterface = reader.read_str_elem(xml)
+        self.assertIsInstance(elem, ar_element.ParameterInterface)
+        self.assertEqual(elem.name, "InterfaceName")
+        self.assertEqual(len(elem.parameters), 1)
+        parameter = elem.parameters[0]
+        self.assertEqual(parameter.name, "Param1")
+        self.assertEqual(str(parameter.type_ref), "AUTOSAR_Platform/ImplementationDataTypes/uint8")
+
+    def test_create_interface_with_two_parameters_using_constructor(self):
+        ref_to_uint8_type = ar_element.AutosarDataTypeRef("AUTOSAR_Platform/ImplementationDataTypes/uint8",
+                                                          ar_enum.IdentifiableSubTypes.IMPLEMENTATION_DATA_TYPE)
+        ref_to_uint16_type = ar_element.AutosarDataTypeRef("AUTOSAR_Platform/ImplementationDataTypes/uint16",
+                                                           ar_enum.IdentifiableSubTypes.IMPLEMENTATION_DATA_TYPE)
+        element = ar_element.ParameterInterface("InterfaceName", [
+            ar_element.ParameterDataPrototype("Param1", type_ref=ref_to_uint8_type),
+            ar_element.ParameterDataPrototype("Param2", type_ref=ref_to_uint16_type)
+        ])
+        writer = autosar.xml.Writer()
+        xml = '''<PARAMETER-INTERFACE>
+  <SHORT-NAME>InterfaceName</SHORT-NAME>
+  <PARAMETERS>
+    <PARAMETER-DATA-PROTOTYPE>
+      <SHORT-NAME>Param1</SHORT-NAME>
+      <TYPE-TREF DEST="IMPLEMENTATION-DATA-TYPE">AUTOSAR_Platform/ImplementationDataTypes/uint8</TYPE-TREF>
+    </PARAMETER-DATA-PROTOTYPE>
+    <PARAMETER-DATA-PROTOTYPE>
+      <SHORT-NAME>Param2</SHORT-NAME>
+      <TYPE-TREF DEST="IMPLEMENTATION-DATA-TYPE">AUTOSAR_Platform/ImplementationDataTypes/uint16</TYPE-TREF>
+    </PARAMETER-DATA-PROTOTYPE>
+  </PARAMETERS>
+</PARAMETER-INTERFACE>'''
+        self.assertEqual(writer.write_str_elem(element), xml)
+        reader = autosar.xml.Reader()
+        elem: ar_element.ParameterInterface = reader.read_str_elem(xml)
+        self.assertIsInstance(elem, ar_element.ParameterInterface)
+        self.assertEqual(elem.name, "InterfaceName")
+        self.assertEqual(len(elem.parameters), 2)
+        parameter = elem.parameters[0]
+        self.assertEqual(parameter.name, "Param1")
+        self.assertEqual(str(parameter.type_ref), "AUTOSAR_Platform/ImplementationDataTypes/uint8")
+        parameter = elem.parameters[1]
+        self.assertEqual(parameter.name, "Param2")
+        self.assertEqual(str(parameter.type_ref), "AUTOSAR_Platform/ImplementationDataTypes/uint16")
+
+    def test_create_interface_with_one_parameter_using_make_method(self):
+        ref_to_uint8_type = ar_element.AutosarDataTypeRef("AUTOSAR_Platform/ImplementationDataTypes/uint8",
+                                                          ar_enum.IdentifiableSubTypes.IMPLEMENTATION_DATA_TYPE)
+        element = ar_element.ParameterInterface("InterfaceName")
+        element.make_parameter("Param1", type_ref=ref_to_uint8_type)
+        writer = autosar.xml.Writer()
+        xml = '''<PARAMETER-INTERFACE>
+  <SHORT-NAME>InterfaceName</SHORT-NAME>
+  <PARAMETERS>
+    <PARAMETER-DATA-PROTOTYPE>
+      <SHORT-NAME>Param1</SHORT-NAME>
+      <TYPE-TREF DEST="IMPLEMENTATION-DATA-TYPE">AUTOSAR_Platform/ImplementationDataTypes/uint8</TYPE-TREF>
+    </PARAMETER-DATA-PROTOTYPE>
+  </PARAMETERS>
+</PARAMETER-INTERFACE>'''
+        self.assertEqual(writer.write_str_elem(element), xml)
+        reader = autosar.xml.Reader()
+        elem: ar_element.ParameterInterface = reader.read_str_elem(xml)
+        self.assertIsInstance(elem, ar_element.ParameterInterface)
+        self.assertEqual(elem.name, "InterfaceName")
+        self.assertEqual(len(elem.parameters), 1)
+        parameter = elem.parameters[0]
+        self.assertEqual(parameter.name, "Param1")
+        self.assertEqual(str(parameter.type_ref), "AUTOSAR_Platform/ImplementationDataTypes/uint8")
+
+    def test_create_interface_with_two_parameters_using_make_method(self):
+        ref_to_uint8_type = ar_element.AutosarDataTypeRef("AUTOSAR_Platform/ImplementationDataTypes/uint8",
+                                                          ar_enum.IdentifiableSubTypes.IMPLEMENTATION_DATA_TYPE)
+        ref_to_uint16_type = ar_element.AutosarDataTypeRef("AUTOSAR_Platform/ImplementationDataTypes/uint16",
+                                                           ar_enum.IdentifiableSubTypes.IMPLEMENTATION_DATA_TYPE)
+        element = ar_element.ParameterInterface("InterfaceName")
+        element.make_parameter("Param1", type_ref=ref_to_uint8_type)
+        element.make_parameter("Param2", type_ref=ref_to_uint16_type)
+
+        writer = autosar.xml.Writer()
+        xml = '''<PARAMETER-INTERFACE>
+  <SHORT-NAME>InterfaceName</SHORT-NAME>
+  <PARAMETERS>
+    <PARAMETER-DATA-PROTOTYPE>
+      <SHORT-NAME>Param1</SHORT-NAME>
+      <TYPE-TREF DEST="IMPLEMENTATION-DATA-TYPE">AUTOSAR_Platform/ImplementationDataTypes/uint8</TYPE-TREF>
+    </PARAMETER-DATA-PROTOTYPE>
+    <PARAMETER-DATA-PROTOTYPE>
+      <SHORT-NAME>Param2</SHORT-NAME>
+      <TYPE-TREF DEST="IMPLEMENTATION-DATA-TYPE">AUTOSAR_Platform/ImplementationDataTypes/uint16</TYPE-TREF>
+    </PARAMETER-DATA-PROTOTYPE>
+  </PARAMETERS>
+</PARAMETER-INTERFACE>'''
+        self.assertEqual(writer.write_str_elem(element), xml)
+        reader = autosar.xml.Reader()
+        elem: ar_element.ParameterInterface = reader.read_str_elem(xml)
+        self.assertIsInstance(elem, ar_element.ParameterInterface)
+        self.assertEqual(elem.name, "InterfaceName")
+        self.assertEqual(len(elem.parameters), 2)
+        parameter = elem.parameters[0]
+        self.assertEqual(parameter.name, "Param1")
+        self.assertEqual(str(parameter.type_ref), "AUTOSAR_Platform/ImplementationDataTypes/uint8")
+        parameter = elem.parameters[1]
+        self.assertEqual(parameter.name, "Param2")
+        self.assertEqual(str(parameter.type_ref), "AUTOSAR_Platform/ImplementationDataTypes/uint16")
 
 
 if __name__ == '__main__':